ai-agents/requirements.txt
openai
python-dotenv
crewai[tools]
openai-agents
streamlit
